requirements.txt
python-telegram-bot==20.7
supabase==2.3.0
python-dotenv==1.0.0
httpx==0.24.0